s3: libsmb: Make cli_list_old() take 32-bit attribute.
Only uses 16 on the wire. No change needed in callers.

+++ source3/libsmb/proto.h
@@ -755,7 +755,7 @@ NTSTATUS cli_posix_whoami(struct cli_state *cli,
 NTSTATUS is_bad_finfo_name(const struct cli_state *cli,
 			const struct file_info *finfo);
 
-NTSTATUS cli_list_old(struct cli_state *cli,const char *Mask,uint16_t attribute,
+NTSTATUS cli_list_old(struct cli_state *cli,const char *Mask,uint32_t attribute,
 		      NTSTATUS (*fn)(const char *, struct file_info *,
 				 const char *, void *), void *state);
 NTSTATUS cli_list_trans(struct cli_state *cli, const char *mask,
